refactor(bridge): split private_ai_provider_verifier.py into a package
Break the 1214-line scripts/private_ai_provider_verifier.py into a
`scripts/provider_verifier/` package:

- common.py       — shared helpers (emit, hashing, evidence bundles, ...)
- tinfoil.py      — verify_tinfoil
- nearai.py       — verify_nearai
- phala_direct.py — verify_phala_direct (+ compose-hash helper)
- chutes.py       — verify_chutes + all chutes_* helpers
- __init__.py     — re-exports the four verify_* entry points

scripts/private_ai_provider_verifier.py is now a thin shim that preserves
both contracts: the CLI entry (`uv run python scripts/private_ai_provider_verifier.py`,
spawned from src/aci/verifier/external.rs) and the importable surface
(`import private_ai_provider_verifier as bridge; bridge.verify_phala_direct`,
used by scripts/soundness_phala_direct.py). confidential_verifier imports stay
lazy/in-function so the package imports without the vendored deps on sys.path.

Pure relocation: no logic changes. Verified: python3 -m compileall scripts,
import-surface smoke test, `uv run scripts/soundness_phala_direct.py` (OK), and
cargo test contract_verifier_bridge / phala_direct_bridge / smoke_scripts (all pass).
