+++ scripts/provider_verifier/__init__.py
@@ -1,0 +1,8 @@
+"""Provider-verifier bridge package (split from private_ai_provider_verifier.py)."""
+
+from .chutes import verify_chutes
+from .nearai import verify_nearai
+from .phala_direct import verify_phala_direct
+from .tinfoil import verify_tinfoil
+
+__all__ = ["verify_chutes", "verify_nearai", "verify_phala_direct", "verify_tinfoil"]
